refactor: View와 Validator를 각각 파일별로 나눠줌
