+++ src/View/OutputView.js
@@ -23,11 +23,7 @@ const OutputView = {
   startOutput() {
     Console.print(START_OUTPUT);
   },
-  /**
-   * 현재까지 이동한 다리의 상태를 정해진 형식에 맞춰 출력한다.
-   * <p>
-   * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
-   */
+
   printMap(bridge, location, moveSuccess) {
     const map = [EMPTY_STRING, EMPTY_STRING];
     map[UP_INDEX] += this.makeOneLineMap(
@@ -67,11 +63,6 @@ const OutputView = {
     return X_CHAR;
   },
 
-  /**
-   * 게임의 최종 결과를 정해진 형식에 맞춰 출력한다.
-   * <p>
-   * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
-   */
   printResult(bridge, location, isSuccess, attempts) {
     Console.print(FINAL_GAME_RESULT);
     this.printMap(bridge, location, isSuccess);
